24. Authorization with policies

+++ app/Http/Controllers/ThreadsController.php
@@ -85,7 +85,11 @@ class ThreadsController extends Controller
 
     public function destroy($channel, Thread $thread)
     {
-//        $thread->replies()->delete();
+//        if ($thread->user_id != auth()->id()) {
+//            abort(403, 'You do not have permission to do this.');
+//        }
+        $this->authorize('update', $thread);
+
         $thread->delete();
 
         if (request()->wantsJson()) {
